[EXPERIMENTAL] Server codebase refactoring
